+++ RELEASES.md
@@ -14,6 +14,7 @@ Upcoming
 
 - Move 'Dynamic parameters' from showConfiguration to showEnvironmentVariables
 - Show provided JmsDestinations with usage in showSecurityItems
+- Introduction of the HashPipe